requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
asyncpg
asyncpixel==1.0.8
asyncpraw==7.2.0
asyncprawcore==2.0.1
attrs==20.3.0
braceexpand==0.1.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
discord-py-slash-command==1.1.2
discord.py==1.7.1
DiscordUtils==1.2.9.2
docopt==0.6.2
humanize
idna==2.10
import-expression==1.1.4
importlib-metadata==3.10.1
jishaku==1.20.0.220
kiwisolver==1.3.1
matplotlib==3.4.1
MojangAPI==0.0.5
multidict==5.1.0
num2words==0.5.10
numpy==1.20.2
olefile==0.46
Pillow
pycparser==2.20
pydantic==1.8.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
ReactivePy==2.0.0.dev0
requests==2.25.1
six==1.15.0
typing-extensions
update-checker==0.18.0
urllib3==1.26.4
yarl==1.6.3
youtube-dl==2021.4.26
zipp==3.4.1
scipy